chore: remove legacy store utils
