+++ client/src/lib/store/constants.ts
@@ -1,0 +1,15 @@
+export const SPEAKER_COLORS = [
+  "hsl(217, 91%, 48%)",
+  "hsl(142, 76%, 36%)",
+  "hsl(271, 81%, 48%)",
+  "hsl(43, 96%, 42%)",
+  "hsl(340, 82%, 52%)",
+  "hsl(190, 90%, 40%)",
+  "hsl(25, 95%, 53%)",
+  "hsl(300, 76%, 45%)",
+];
+
+export const MAX_HISTORY = 100;
+export const PERSIST_THROTTLE_MS = 500;
+export const PLAYING_TIME_PERSIST_STEP = 1;
+export const PAUSED_TIME_PERSIST_STEP = 0.25;
