+++ client/src/lib/store/utils/id.ts
@@ -1,0 +1,1 @@
+export const generateId = () => crypto.randomUUID();
